Format custom typings and split them into separate files
