+++ custom_typings/command-line-args.d.ts
@@ -1,0 +1,23 @@
+declare module 'command-line-args' {
+  interface ArgDescriptor {
+    name: string;
+    // type: Object;
+    alias?: string;
+    description: string;
+    defaultValue?: any;
+    type: (val: string) => any;
+    multiple?: boolean;
+  }
+  interface UsageOpts {
+    title: string;
+    header: string;
+  }
+  interface CLI {
+    parse(): any;
+    getUsage(opts: UsageOpts): string;
+  }
+  function commandLineArgs(args: ArgDescriptor[]): CLI;
+  module commandLineArgs {}
+
+  export = commandLineArgs;
+}
