+++ custom_typings/escodegen.d.ts
@@ -1,0 +1,13 @@
+declare module 'escodegen' {
+  interface GenerateOpts {
+    comment?: boolean;
+    format?: {
+      indent?: {
+        style?: string;
+        base?: number;
+        adjustMultilineComment: boolean;
+      }
+    }
+  }
+  export function generate(ast: any, opts?: GenerateOpts): string;
+}
