+++ custom_typings/espree.d.ts
@@ -1,0 +1,6 @@
+declare module 'espree' {
+  interface ParseOpts {
+    attachComment: boolean;
+  }
+  export function parse(text: string, opts?: ParseOpts): any;
+}
