+++ custom_typings/estree-walker.d.ts
@@ -1,0 +1,3 @@
+declare module 'estree-walker' {
+  export function walk(n: any, callbacks: {enter: (node: any) => any}): void;
+}
