+++ tsconfig.json
@@ -16,7 +16,7 @@
         "**/*.tsx",
         "!node_modules/**",
         "typings/tsd.d.ts",
-        "custom_typings/main.d.ts"
+        "custom_typings/*.d.ts"
     ],
     "files": [
         "cleanup-passes/bower.ts",
@@ -26,7 +26,17 @@
         "cleanup-passes/travis.ts",
         "cleanup-passes/util.ts",
         "cleanup.ts",
+        "custom_typings/command-line-args.d.ts",
+        "custom_typings/dom5.d.ts",
+        "custom_typings/escodegen.d.ts",
+        "custom_typings/espree.d.ts",
+        "custom_typings/estree-walker.d.ts",
+        "custom_typings/github-cache.d.ts",
+        "custom_typings/hydrolysis.d.ts",
         "custom_typings/main.d.ts",
+        "custom_typings/nodegit.d.ts",
+        "custom_typings/pad.d.ts",
+        "custom_typings/promisify-node.d.ts",
         "element-repo.ts",
         "tedium.ts",
         "typings/js-yaml/js-yaml.d.ts",
